Put str_id macro into a separate module
For better modularity.
