+++ src/lib.rs
@@ -93,6 +93,9 @@ extern crate uuid;
 /// Range macros
 #[macro_use]
 mod range_macros;
+/// ID macros
+#[macro_use]
+mod id_macros;
 /// shared constants
 mod consts;
 /// rust definitions of ioctl structs and consts
